added landing fragment instead of just activity. Added dynamic inflation
