+++ app/src/main/res/layout/landing_activity.xml
@@ -1,53 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.LandingActivity"
     android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/landingHeader"
-        android:layout_width="wrap_content"
-        android:layout_height="52dp"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="20dp"
-        android:layout_marginBottom="10dp"
-        android:gravity="center"
-        android:includeFontPadding="false"
-        android:text="@string/app_name"
-        android:textColor="#000"
-        android:textSize="25sp"
-        android:textStyle="bold"
-        android:layout_gravity="center"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/goToQuizButton"
+    <fragment
+        android:name="com.example.city_matcher.UI.LandingFragment"
+        android:id="@+id/landingFragment"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/quizButtonPrompt"/>
-
-    <Button
-        android:id="@+id/showMatchedCityButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/currentMatchedCityPrompt" />
-
-    <Button
-        android:id="@+id/logoutButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/logoutButtonPrompt" />
-
-    <Button
-        android:id="@+id/deleteAccountButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/deleteButtonPrompt" />
-
+        android:layout_height="match_parent"/>
 </LinearLayout>
